// src/components/forms/wizards/RegCardWizard/pages/Privileges/sections/Disability/types.ts
export const DROPDOWN_LABEL = 'Инвалидность';

export enum LABELS {
  SOMAT = 'Сомат',
  SET_DATE = 'Дата установления',
  TIME = 'Бессрочно',
  GROUP = 'Группа',
  NEXT_PLAN = 'Очередное переосв.',
  WORK_PLACE = 'Место работы',
  DEGREE = 'Степень утраты трудосп.',
  ORIGIN = 'Перв.',
  HOSPITAL = 'Стационар',
  NOTE = 'Примечание',
}